+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/ChatClient.kt
@@ -167,8 +167,11 @@ import io.getstream.logging.SilentStreamLogger
 import io.getstream.logging.StreamLog
 import io.getstream.logging.android.AndroidStreamLogger
 import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.cancelChildren
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.withTimeoutOrNull
 import okhttp3.OkHttpClient
 import java.io.File
@@ -1071,6 +1074,7 @@ internal constructor(
             lifecycleObserver.dispose()
             appSettingsManager.clear()
             _repositoryFacade = null
+            postponeCancelScope()
             Result.success(Unit)
         }
 
@@ -1087,6 +1091,17 @@ internal constructor(
     @WorkerThread
     public fun disconnect() {
         disconnect(true).execute()
+    }
+
+    /**
+     * Cancel all jobs on the ChatClient Scope.
+     * This method postpone a new coroutine to cancel all children jobs.
+     */
+    private fun postponeCancelScope() {
+        scope.launch {
+            delay(DELAY_TIME_TO_CANCEL_CHILDREN)
+            scope.coroutineContext.cancelChildren()
+        }
     }
 
     //region: api calls
@@ -2810,6 +2825,7 @@ internal constructor(
         @JvmStatic
         public var OFFLINE_SUPPORT_ENABLED: Boolean = false
 
+        private const val DELAY_TIME_TO_CANCEL_CHILDREN = 100L
         private const val MAX_COOLDOWN_TIME_SECONDS = 120
         private const val KEY_MESSAGE_ACTION = "image_action"
         private const val MESSAGE_ACTION_SEND = "send"