+++ stream-chat-android-ui-components/src/main/kotlin/com/getstream/sdk/chat/viewmodel/MessageInputViewModel.kt
@@ -125,7 +125,8 @@ public class MessageInputViewModel @JvmOverloads constructor(
      */
     public var typingUpdatesBuffer: TypingUpdatesBuffer = DefaultTypingUpdatesBuffer(
         onTypingStarted = ::keystroke,
-        onTypingStopped = ::stopTyping
+        onTypingStopped = ::stopTyping,
+        coroutineScope = viewModelScope
     )
 
     /**