+++ stream-chat-android-ui-components/src/main/kotlin/com/getstream/sdk/chat/viewmodel/messages/MessageListViewModel.kt
@@ -526,6 +526,64 @@ public class MessageListViewModel(
                     }
                 )
             }
+            is Event.BanUser -> {
+                val channelClient = chatClient.channel(cid)
+                channelClient.banUser(
+                    targetId = event.user.id,
+                    reason = event.reason,
+                    timeout = event.timeout,
+                ).enqueue(
+                    onError = { chatError ->
+                        val errorMessage = chatError.message ?: chatError.cause?.message ?: "Unable to ban the user"
+                        logger.e { errorMessage }
+
+                        _errorEvents.postValue(EventWrapper(ErrorEvent.BlockUserError(chatError)))
+                    }
+                )
+            }
+            is Event.UnbanUser -> {
+                val channelClient = chatClient.channel(cid)
+                channelClient.unbanUser(
+                    targetId = event.user.id,
+                ).enqueue(
+                    onError = { chatError ->
+                        val errorMessage = chatError.message ?: chatError.cause?.message ?: "Unable to unban the user"
+                        logger.e { errorMessage }
+
+                        _errorEvents.postValue(EventWrapper(ErrorEvent.BlockUserError(chatError)))
+                    }
+                )
+            }
+            is Event.ShadowBanUser -> {
+                val channelClient = chatClient.channel(cid)
+                channelClient.shadowBanUser(
+                    targetId = event.user.id,
+                    reason = event.reason,
+                    timeout = event.timeout,
+                ).enqueue(
+                    onError = { chatError ->
+                        val errorMessage = chatError.message ?: chatError.cause?.message
+                            ?: "Unable to shadow ban the user"
+                        logger.e { errorMessage }
+
+                        _errorEvents.postValue(EventWrapper(ErrorEvent.BlockUserError(chatError)))
+                    }
+                )
+            }
+            is Event.RemoveShadowBanFromUser -> {
+                val channelClient = chatClient.channel(cid)
+                channelClient.removeShadowBan(
+                    targetId = event.user.id,
+                ).enqueue(
+                    onError = { chatError ->
+                        val errorMessage =
+                            chatError.message ?: chatError.cause?.message ?: "Unable to remove the user shadow ban"
+                        logger.e { errorMessage }
+
+                        _errorEvents.postValue(EventWrapper(ErrorEvent.BlockUserError(chatError)))
+                    }
+                )
+            }
             is Event.ReplyMessage -> {
                 chatClient.setMessageForReply(event.cid, event.repliedMessage).enqueue(
                     onError = { chatError ->
@@ -1003,7 +1061,59 @@ public class MessageListViewModel(
          * @param user The user to be blocked.
          * @param cid The full channel id, i.e. "messaging:123".
          */
+        @Deprecated(
+            "Deprecated in order to make the action more explicit." +
+                "Use `MessageListViewModel.ShadowBanUser` if you want to keep the same functionality " +
+                "(shadow banning) or `MessageListViewModel.BanUser` if you want to ban a user.",
+            replaceWith = ReplaceWith(
+                "ShadowBanUser",
+                "package com.getstream.sdk.chat.viewmodel.messages.MessageListViewModel.ShadowBanUser"
+            )
+        )
         public data class BlockUser(val user: User, val cid: String) : Event()
+
+        /**
+         * When the user bans another user.
+         *
+         * @param user The user to be blocked.
+         * @param reason The reason for banning the user.
+         * @param timeout The time until the user is automatically unbanned, expressed in minutes.
+         * A null value represent an indefinite ban.
+         */
+        public data class BanUser(
+            val user: User,
+            val reason: String? = null,
+            val timeout: Int? = null,
+        ) : Event()
+
+        /**
+         * When the user unbans another user.
+         *
+         * @param user The user to be blocked.
+         */
+        public data class UnbanUser(val user: User) : Event()
+
+        /**
+         * When the user shadow bans another user.
+         *
+         * @param user The user to be blocked.
+         * @param reason The reason for shadow banning the user.
+         * @param timeout The time until the shadow ban is automatically removed
+         * for the given user, expressed in minutes.
+         * A null value represent an indefinite ban.
+         */
+        public data class ShadowBanUser(
+            val user: User,
+            val reason: String? = null,
+            val timeout: Int? = null,
+        ) : Event()
+
+        /**
+         * When the user removes the shadow ban from another user.
+         *
+         * @param user The user to be blocked.
+         */
+        public data class RemoveShadowBanFromUser(val user: User) : Event()
 
         /**
          * When the user replies to a message.